chore: reorganize sites code
