+++ app/components/Nav/Main/MainNavigator.js
@@ -52,7 +52,6 @@ import ContactForm from '../../Views/Settings/Contacts/ContactForm';
 import ActivityView from '../../Views/ActivityView';
 import RewardsNavigator from '../../UI/Rewards/RewardsNavigator';
 import TrendingView from '../../Views/TrendingView/TrendingView';
-import SitesListView from '../../Views/TrendingView/SitesListView';
 import SwapsAmountView from '../../UI/Swaps';
 import SwapsQuotesView from '../../UI/Swaps/QuotesView';
 import CollectiblesDetails from '../../UI/CollectibleModal';
@@ -133,6 +132,7 @@ import {
   TOKEN,
 } from '../../Views/AddAsset/AddAsset.constants';
 import { strings } from '../../../../locales/i18n';
+import SitesFullView from '../../Views/SitesFullView/SitesFullView';
 
 const Stack = createStackNavigator();
 const Tab = createBottomTabNavigator();
@@ -293,8 +293,8 @@ const TrendingHome = () => (
       options={{ headerShown: false }}
     />
     <Stack.Screen
-      name={Routes.SITES_LIST_VIEW}
-      component={SitesListView}
+      name={Routes.SITES_FULL_VIEW}
+      component={SitesFullView}
       options={{
         headerShown: false,
         animationEnabled: true,
