--- app/components/Views/TrendingView/SectionSites/SiteRowItem/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SiteRowItem';
-export type { SiteData } from './SiteRowItem';
