--- app/components/Views/TrendingView/SectionSites/SiteSkeleton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SiteSkeleton';
