--- app/components/Views/TrendingView/SectionSites/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useSitesData } from './useSitesData';
