--- app/components/Views/TrendingView/SectionSites/index.ts
@@ -1,5 +1,0 @@
-export { default as SiteRowItem } from './SiteRowItem';
-export { default as SiteRowItemWrapper } from './SiteRowItemWrapper';
-export { default as SiteSkeleton } from './SiteSkeleton';
-export { useSitesData } from './hooks/useSitesData';
-export type { SiteData } from './SiteRowItem';
