--- app/components/Views/TrendingView/SitesListView/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SitesListView';
